requirements.txt
mcp
faiss-cpu
sentence-transformers
openai